requirements.txt
# QuickSell Chatbot Dependencies
# Python 3.13+ (Windows)

# Web Framework
Flask==3.0.0
Werkzeug==3.0.1

# HTTP Requests
requests==2.31.0
urllib3==2.1.0

# Database
mysql-connector-python==8.2.0
PyMySQL==1.1.0

# Environment Variables
python-dotenv==1.0.0

# Production Server (Optional but recommended)
gunicorn==21.2.0
gevent==23.9.1

# Utilities
python-dateutil==2.8.2

# Excel Processing
pandas==2.1.4
openpyxl==3.1.2

# Image Processing
Pillow==10.1.0

# Security
cryptography==41.0.7
bcrypt==5.0.0

# Logging & Monitoring (Optional)
colorlog==6.8.0

# Testing (Optional)
pytest==7.4.3
pytest-cov==4.1.0

# Code Quality (Optional)
flake8==6.1.0
black==23.12.1
